@@ -2,7 +2,7 @@
   "name": "vinyl-fs",
   "version": "3.0.3",
   "description": "Vinyl adapter for the file system.",
-  "author": "Gulp Team <team@gulpjs.com> (http://gulpjs.com/)",
+  "author": "Gulp Team <team@gulpjs.com> (https://gulpjs.com/)",
   "contributors": [
     "Eric Schoffstall <yo@contra.io>",
     "Blaine Bublitz <blaine.bublitz@gmail.com>"
@@ -10,7 +10,7 @@
   "repository": "gulpjs/vinyl-fs",
   "license": "MIT",
   "engines": {
-    "node": ">= 0.10"
+    "node": ">=10.13.0"
   },
   "main": "index.js",
   "files": [
@@ -19,43 +19,45 @@
     "lib"
   ],
   "scripts": {
-    "lint": "eslint . && jscs index.js lib/ test/",
+    "lint": "eslint .",
     "pretest": "npm run lint",
-    "test": "mocha --async-only",
-    "cover": "istanbul cover _mocha --report lcovonly",
-    "coveralls": "npm run cover && istanbul-coveralls"
+    "test": "nyc mocha --async-only"
   },
   "dependencies": {
-    "fs-mkdirp-stream": "^1.0.0",
-    "glob-stream": "^6.1.0",
-    "graceful-fs": "^4.0.0",
-    "iconv-lite": "^0.4.24",
+    "fs-mkdirp-stream": "^2.0.1",
+    "glob-stream": "^8.0.0",
+    "graceful-fs": "^4.2.11",
+    "iconv-lite": "^0.6.3",
     "is-valid-glob": "^1.0.0",
-    "lazystream": "^1.0.0",
-    "lead": "^1.0.0",
-    "object.assign": "^4.0.4",
-    "pumpify": "^1.3.5",
-    "readable-stream": "^2.3.3",
-    "remove-bom-buffer": "^3.0.0",
-    "remove-bom-stream": "^1.2.0",
-    "resolve-options": "^1.1.0",
-    "through2": "^2.0.0",
-    "to-through": "^2.0.0",
-    "value-or-function": "^3.0.0",
-    "vinyl": "^2.0.0",
-    "vinyl-sourcemap": "^1.1.0"
+    "lazystream": "^1.0.1",
+    "lead": "^4.0.0",
+    "normalize-path": "3.0.0",
+    "resolve-options": "^2.0.0",
+    "stream-composer": "^1.0.2",
+    "streamx": "^2.14.0",
+    "to-through": "^3.0.0",
+    "value-or-function": "^4.0.0",
+    "vinyl": "^3.0.0",
+    "vinyl-sourcemap": "^2.0.0"
   },
   "devDependencies": {
-    "eslint": "^1.10.3",
-    "eslint-config-gulp": "^2.0.0",
-    "expect": "^1.19.0",
-    "istanbul": "^0.4.3",
-    "istanbul-coveralls": "^1.0.3",
-    "jscs": "^2.4.0",
-    "jscs-preset-gulp": "^1.0.0",
-    "mississippi": "^1.2.0",
-    "mocha": "^3.5.0",
-    "rimraf": "^2.6.1"
+    "eslint": "^7.32.0",
+    "eslint-config-gulp": "^5.0.1",
+    "expect": "^27.5.1",
+    "mocha": "^8.4.0",
+    "nyc": "^15.1.0",
+    "readable-stream": "3.6.0",
+    "rimraf": "^3.0.2",
+    "sinon": "^15.0.2"
+  },
+  "nyc": {
+    "reporter": [
+      "lcov",
+      "text-summary"
+    ]
+  },
+  "prettier": {
+    "singleQuote": true
   },
   "keywords": [
     "gulp",
