@@ -3,17 +3,14 @@
 var fs = require('graceful-fs');
 var File = require('vinyl');
 var expect = require('expect');
-var miss = require('mississippi');
+var sinon = require('sinon');
 
 var vfs = require('../');
 
 var cleanup = require('./utils/cleanup');
 var isWindows = require('./utils/is-windows');
 var testConstants = require('./utils/test-constants');
-
-var from = miss.from;
-var pipe = miss.pipe;
-var concat = miss.concat;
+var describeStreams = require('./utils/suite');
 
 var inputBase = testConstants.inputBase;
 var outputBase = testConstants.outputBase;
@@ -22,23 +19,25 @@ var contents = testConstants.contents;
 
 var clean = cleanup(outputBase);
 
-describe('.dest() with custom owner', function() {
+describeStreams('.dest() with custom owner', function (stream) {
+  var from = stream.Readable.from;
+  var pipeline = stream.pipeline;
 
   beforeEach(clean);
   afterEach(clean);
 
-  it('calls fchown when the uid and/or gid are provided on the vinyl stat', function(done) {
+  it('calls fchown when the uid and/or gid are provided on the vinyl stat', function (done) {
     if (isWindows) {
       this.skip();
       return;
     }
 
-    var fchownSpy = expect.spyOn(fs, 'fchown').andCallThrough();
+    var fchownSpy = sinon.spy(fs, 'fchown');
 
     var file = new File({
       base: inputBase,
       path: inputPath,
-      contents: new Buffer(contents),
+      contents: Buffer.from(contents),
       stat: {
         uid: 1001,
         gid: 1001,
@@ -46,30 +45,27 @@ describe('.dest() with custom owner', function() {
     });
 
     function assert() {
-      expect(fchownSpy.calls.length).toEqual(1);
-      expect(fchownSpy.calls[0].arguments[1]).toEqual(1001);
-      expect(fchownSpy.calls[0].arguments[2]).toEqual(1001);
+      expect(fchownSpy.callCount).toEqual(1);
+      expect(fchownSpy.getCall(0).args[1]).toEqual(1001);
+      expect(fchownSpy.getCall(0).args[2]).toEqual(1001);
+      done();
     }
 
-    pipe([
-      from.obj([file]),
-      vfs.dest(outputBase),
-      concat(assert),
-    ], done);
+    pipeline([from([file]), vfs.dest(outputBase)], assert);
   });
 
-  it('does not call fchown when the uid and gid provided on the vinyl stat are invalid', function(done) {
+  it('does not call fchown when the uid and gid provided on the vinyl stat are invalid', function (done) {
     if (isWindows) {
       this.skip();
       return;
     }
 
-    var fchownSpy = expect.spyOn(fs, 'fchown').andCallThrough();
+    var fchownSpy = sinon.spy(fs, 'fchown');
 
     var file = new File({
       base: inputBase,
       path: inputPath,
-      contents: new Buffer(contents),
+      contents: Buffer.from(contents),
       stat: {
         uid: -1,
         gid: -1,
@@ -77,13 +73,10 @@ describe('.dest() with custom owner', function() {
     });
 
     function assert() {
-      expect(fchownSpy.calls.length).toEqual(0);
+      expect(fchownSpy.callCount).toEqual(0);
+      done();
     }
 
-    pipe([
-      from.obj([file]),
-      vfs.dest(outputBase),
-      concat(assert),
-    ], done);
+    pipeline([from([file]), vfs.dest(outputBase)], assert);
   });
 });
