@@ -3,17 +3,14 @@
 var fs = require('graceful-fs');
 var File = require('vinyl');
 var expect = require('expect');
-var miss = require('mississippi');
+var sinon = require('sinon');
 
 var vfs = require('../');
 
 var cleanup = require('./utils/cleanup');
 var isWindows = require('./utils/is-windows');
 var testConstants = require('./utils/test-constants');
-
-var from = miss.from;
-var pipe = miss.pipe;
-var concat = miss.concat;
+var describeStreams = require('./utils/suite');
 
 var inputBase = testConstants.inputBase;
 var outputBase = testConstants.outputBase;
@@ -23,12 +20,14 @@ var contents = testConstants.contents;
 
 var clean = cleanup(outputBase);
 
-describe('.dest() with custom times', function() {
+describeStreams('.dest() with custom times', function (stream) {
+  var from = stream.Readable.from;
+  var pipeline = stream.pipeline;
 
   beforeEach(clean);
   afterEach(clean);
 
-  it('does not call futimes when no mtime is provided on the vinyl stat', function(done) {
+  it('does not call futimes when no mtime is provided on the vinyl stat', function (done) {
     // Changing the time of a directory errors in Windows.
     // Windows is treated as though it does not have permission to make this operation.
     if (isWindows) {
@@ -38,31 +37,28 @@ describe('.dest() with custom times', function() {
 
     var earlier = Date.now() - 1001;
 
-    var futimesSpy = expect.spyOn(fs, 'futimes').andCallThrough();
+    var futimesSpy = sinon.spy(fs, 'futimes');
 
     var file = new File({
       base: inputBase,
       path: inputPath,
-      contents: new Buffer(contents),
+      contents: Buffer.from(contents),
       stat: {},
     });
 
     function assert() {
       var stats = fs.lstatSync(outputPath);
 
-      expect(futimesSpy.calls.length).toEqual(0);
+      expect(futimesSpy.callCount).toEqual(0);
       expect(stats.atime.getTime()).toBeGreaterThan(earlier);
       expect(stats.mtime.getTime()).toBeGreaterThan(earlier);
+      done();
     }
 
-    pipe([
-      from.obj([file]),
-      vfs.dest(outputBase, { cwd: __dirname }),
-      concat(assert),
-    ], done);
+    pipeline([from([file]), vfs.dest(outputBase, { cwd: __dirname })], assert);
   });
 
-  it('calls futimes when an mtime is provided on the vinyl stat', function(done) {
+  it('calls futimes when an mtime is provided on the vinyl stat', function (done) {
     if (isWindows) {
       this.skip();
       return;
@@ -71,37 +67,34 @@ describe('.dest() with custom times', function() {
     // Use new mtime
     var mtime = new Date(Date.now() - 2048);
 
-    var futimesSpy = expect.spyOn(fs, 'futimes').andCallThrough();
+    var futimesSpy = sinon.spy(fs, 'futimes');
 
     var file = new File({
       base: inputBase,
       path: inputPath,
-      contents: new Buffer(contents),
+      contents: Buffer.from(contents),
       stat: {
         mtime: mtime,
       },
     });
 
     function assert() {
-      expect(futimesSpy.calls.length).toEqual(1);
+      expect(futimesSpy.callCount).toEqual(1);
 
       // Compare args instead of fs.lstats(), since mtime may be drifted in x86 Node.js
-      var mtimeSpy = futimesSpy.calls[0].arguments[2];
+      var mtimeSpy = futimesSpy.getCall(0).args[2];
 
-      expect(mtimeSpy.getTime())
-        .toEqual(mtime.getTime());
+      expect(mtimeSpy.getTime()).toEqual(mtime.getTime());
 
       expect(file.stat.mtime).toEqual(mtime);
+
+      done();
     }
 
-    pipe([
-      from.obj([file]),
-      vfs.dest(outputBase, { cwd: __dirname }),
-      concat(assert),
-    ], done);
+    pipeline([from([file]), vfs.dest(outputBase, { cwd: __dirname })], assert);
   });
 
-  it('does not call futimes when provided mtime on the vinyl stat is invalid', function(done) {
+  it('does not call futimes when provided mtime on the vinyl stat is invalid', function (done) {
     if (isWindows) {
       this.skip();
       return;
@@ -109,12 +102,12 @@ describe('.dest() with custom times', function() {
 
     var earlier = Date.now() - 1001;
 
-    var futimesSpy = expect.spyOn(fs, 'futimes').andCallThrough();
+    var futimesSpy = sinon.spy(fs, 'futimes');
 
     var file = new File({
       base: inputBase,
       path: inputPath,
-      contents: new Buffer(contents),
+      contents: Buffer.from(contents),
       stat: {
         mtime: new Date(undefined),
       },
@@ -123,18 +116,16 @@ describe('.dest() with custom times', function() {
     function assert() {
       var stats = fs.lstatSync(outputPath);
 
-      expect(futimesSpy.calls.length).toEqual(0);
+      expect(futimesSpy.callCount).toEqual(0);
       expect(stats.mtime.getTime()).toBeGreaterThan(earlier);
+
+      done();
     }
 
-    pipe([
-      from.obj([file]),
-      vfs.dest(outputBase, { cwd: __dirname }),
-      concat(assert),
-    ], done);
+    pipeline([from([file]), vfs.dest(outputBase, { cwd: __dirname })], assert);
   });
 
-  it('calls futimes when provided mtime on the vinyl stat is valid but provided atime is invalid', function(done) {
+  it('calls futimes when provided mtime on the vinyl stat is valid but provided atime is invalid', function (done) {
     if (isWindows) {
       this.skip();
       return;
@@ -144,12 +135,12 @@ describe('.dest() with custom times', function() {
     var mtime = new Date(Date.now() - 2048);
     var invalidAtime = new Date(undefined);
 
-    var futimesSpy = expect.spyOn(fs, 'futimes').andCallThrough();
+    var futimesSpy = sinon.spy(fs, 'futimes');
 
     var file = new File({
       base: inputBase,
       path: inputPath,
-      contents: new Buffer(contents),
+      contents: Buffer.from(contents),
       stat: {
         atime: invalidAtime,
         mtime: mtime,
@@ -157,21 +148,19 @@ describe('.dest() with custom times', function() {
     });
 
     function assert() {
-      expect(futimesSpy.calls.length).toEqual(1);
+      expect(futimesSpy.callCount).toEqual(1);
 
-      var mtimeSpy = futimesSpy.calls[0].arguments[2];
+      var mtimeSpy = futimesSpy.getCall(0).args[2];
 
       expect(mtimeSpy.getTime()).toEqual(mtime.getTime());
+
+      done();
     }
 
-    pipe([
-      from.obj([file]),
-      vfs.dest(outputBase, { cwd: __dirname }),
-      concat(assert),
-    ], done);
+    pipeline([from([file]), vfs.dest(outputBase, { cwd: __dirname })], assert);
   });
 
-  it('writes file atime and mtime using the vinyl stat', function(done) {
+  it('writes file atime and mtime using the vinyl stat', function (done) {
     if (isWindows) {
       this.skip();
       return;
@@ -181,12 +170,12 @@ describe('.dest() with custom times', function() {
     var atime = new Date(Date.now() - 2048);
     var mtime = new Date(Date.now() - 1024);
 
-    var futimesSpy = expect.spyOn(fs, 'futimes').andCallThrough();
+    var futimesSpy = sinon.spy(fs, 'futimes');
 
     var file = new File({
       base: inputBase,
       path: inputPath,
-      contents: new Buffer(contents),
+      contents: Buffer.from(contents),
       stat: {
         atime: atime,
         mtime: mtime,
@@ -194,21 +183,19 @@ describe('.dest() with custom times', function() {
     });
 
     function assert() {
-      expect(futimesSpy.calls.length).toEqual(1);
+      expect(futimesSpy.callCount).toEqual(1);
 
-      var atimeSpy = futimesSpy.calls[0].arguments[1];
-      var mtimeSpy = futimesSpy.calls[0].arguments[2];
+      var atimeSpy = futimesSpy.getCall(0).args[1];
+      var mtimeSpy = futimesSpy.getCall(0).args[2];
 
       expect(atimeSpy.getTime()).toEqual(atime.getTime());
       expect(mtimeSpy.getTime()).toEqual(mtime.getTime());
       expect(file.stat.mtime).toEqual(mtime);
       expect(file.stat.atime).toEqual(atime);
-    };
 
-    pipe([
-      from.obj([file]),
-      vfs.dest(outputBase, { cwd: __dirname }),
-      concat(assert),
-    ], done);
+      done();
+    }
+
+    pipeline([from([file]), vfs.dest(outputBase, { cwd: __dirname })], assert);
   });
 });
