@@ -5,16 +5,13 @@ var path = require('path');
 var fs = require('graceful-fs');
 var File = require('vinyl');
 var expect = require('expect');
-var miss = require('mississippi');
+var sinon = require('sinon');
 
 var vfs = require('../');
 
+var testStreams = require('./utils/test-streams');
 var testConstants = require('./utils/test-constants');
-
-var pipe = miss.pipe;
-var from = miss.from;
-var concat = miss.concat;
-var through = miss.through;
+var describeStreams = require('./utils/suite');
 
 var inputBase = testConstants.inputBase;
 var inputPath = testConstants.inputPath;
@@ -30,92 +27,95 @@ var encodedContents = testConstants.encodedContents;
 var bomContents = testConstants.bomContents;
 var contents = testConstants.contents;
 
-describe('.src()', function() {
+describeStreams('.src()', function (stream) {
+  var from = stream.Readable.from;
+  var pipeline = stream.pipeline;
 
-  it('throws on invalid glob (empty)', function(done) {
+  var streamUtils = testStreams(stream);
+  var concatArray = streamUtils.concatArray;
+  var compareContents = streamUtils.compareContents;
+
+  it('throws on invalid glob (empty)', function (done) {
     var stream;
     try {
       stream = vfs.src();
     } catch (err) {
-      expect(err).toExist();
-      expect(stream).toNotExist();
+      expect(err).toEqual(expect.anything());
+      expect(stream).not.toEqual(expect.anything());
       done();
     }
   });
 
-  it('throws on invalid glob (empty string)', function(done) {
+  it('throws on invalid glob (empty string)', function (done) {
     var stream;
     try {
       stream = vfs.src('');
     } catch (err) {
-      expect(err).toExist();
-      expect(stream).toNotExist();
+      expect(err).toEqual(expect.anything());
+      expect(stream).not.toEqual(expect.anything());
       done();
     }
   });
 
-  it('throws on invalid glob (number)', function(done) {
+  it('throws on invalid glob (number)', function (done) {
     var stream;
     try {
       stream = vfs.src(123);
     } catch (err) {
-      expect(err).toExist();
-      expect(stream).toNotExist();
+      expect(err).toEqual(expect.anything());
+      expect(stream).not.toEqual(expect.anything());
       done();
     }
   });
 
-  it('throws on invalid glob (nested array)', function(done) {
+  it('throws on invalid glob (nested array)', function (done) {
     var stream;
     try {
       stream = vfs.src([['./fixtures/*.coffee']]);
     } catch (err) {
-      expect(err).toExist();
-      expect(stream).toNotExist();
-      expect(err.message).toInclude('Invalid glob argument');
+      expect(err).toEqual(expect.anything());
+      expect(stream).not.toEqual(expect.anything());
+      expect(err.message).toMatch('Invalid glob argument');
       done();
     }
   });
 
-  it('throws on invalid glob (empty string in array)', function(done) {
+  it('throws on invalid glob (empty string in array)', function (done) {
     var stream;
     try {
       stream = vfs.src(['']);
     } catch (err) {
-      expect(err).toExist();
-      expect(stream).toNotExist();
+      expect(err).toEqual(expect.anything());
+      expect(stream).not.toEqual(expect.anything());
       done();
     }
   });
 
-  it('throws on invalid glob (empty array)', function(done) {
+  it('throws on invalid glob (empty array)', function (done) {
     var stream;
     try {
       stream = vfs.src([]);
     } catch (err) {
-      expect(err).toExist();
-      expect(stream).toNotExist();
+      expect(err).toEqual(expect.anything());
+      expect(stream).not.toEqual(expect.anything());
       done();
     }
   });
 
-  it('emits an error on file not existing', function(done) {
+  it('emits an error on file not existing', function (done) {
     function assert(err) {
-      expect(err).toExist();
+      expect(err).toEqual(expect.anything());
       done();
     }
 
-    pipe([
-      vfs.src('./fixtures/noexist.coffee'),
-      concat(),
-    ], assert);
+    pipeline([vfs.src('./fixtures/noexist.coffee'), concatArray()], assert);
   });
 
-  it('passes through writes', function(done) {
+  it('passes through writes', function (done) {
     var file = new File({
       base: inputBase,
       path: inputPath,
-      contents: new Buffer(contents),
+      contents: Buffer.from(contents),
       stat: fs.statSync(inputPath),
     });
 
@@ -128,40 +128,25 @@ describe('.src()', function() {
 
     srcStream.write(file);
 
-    pipe([
-      srcStream,
-      concat(assert),
-    ], done);
+    pipeline([srcStream, concatArray(assert)], done);
   });
 
-  it('removes BOM from utf8-encoded files by default (buffer)', function(done) {
-    var expectedContent = new Buffer(bomContents.replace('X', '8'));
+  it('removes BOM from utf8-encoded files by default (buffer)', function (done) {
+    var expectedContent = Buffer.from(bomContents.replace('X', '8'));
 
     function assert(files) {
       expect(files.length).toEqual(1);
-      expect(files[0].contents).toMatch(expectedContent);
+      expect(files[0].contents).toEqual(expectedContent);
     }
 
-    pipe([
-      vfs.src(bomInputPath),
-      concat(assert),
-    ], done);
+    pipeline([vfs.src(bomInputPath), concatArray(assert)], done);
   });
 
-  it('removes BOM from utf8-encoded files by default (stream)', function(done) {
-    var expectedContent = new Buffer(bomContents.replace('X', '8'));
+  it('removes BOM from utf8-encoded files by default (stream)', function (done) {
+    var expectedContent = Buffer.from(bomContents.replace('X', '8'));
 
     function assertContent(contents) {
-      expect(contents).toMatch(expectedContent);
-    }
-
-    function compareContents(file, enc, cb) {
-      pipe([
-        file.contents,
-        concat(assertContent),
-      ], function(err) {
-        cb(err, file);
-      });
+      expect(contents).toEqual(expectedContent);
     }
 
     function assert(files) {
@@ -169,41 +154,35 @@ describe('.src()', function() {
       expect(files[0].isStream()).toEqual(true);
     }
 
-    pipe([
-      vfs.src(bomInputPath, { buffer: false }),
-      through.obj(compareContents),
-      concat(assert),
-    ], done);
+    pipeline(
+      [
+        vfs.src(bomInputPath, { buffer: false }),
+        compareContents(assertContent),
+        concatArray(assert),
+      ],
+      done
+    );
   });
 
-  it('does not remove BOM from utf8-encoded files if option is false (buffer)', function(done) {
-    var expectedContent = new Buffer('\ufeff' + bomContents.replace('X', '8'));
+  it('does not remove BOM from utf8-encoded files if option is false (buffer)', function (done) {
+    var expectedContent = Buffer.from('\ufeff' + bomContents.replace('X', '8'));
 
     function assert(files) {
       expect(files.length).toEqual(1);
-      expect(files[0].contents).toMatch(expectedContent);
+      expect(files[0].contents).toEqual(expectedContent);
     }
 
-    pipe([
-      vfs.src(bomInputPath, { removeBOM: false }),
-      concat(assert),
-    ], done);
+    pipeline(
+      [vfs.src(bomInputPath, { removeBOM: false }), concatArray(assert)],
+      done
+    );
   });
 
-  it('does not remove BOM from utf8-encoded files if option is false (stream)', function(done) {
-    var expectedContent = new Buffer('\ufeff' + bomContents.replace('X', '8'));
+  it('does not remove BOM from utf8-encoded files if option is false (stream)', function (done) {
+    var expectedContent = Buffer.from('\ufeff' + bomContents.replace('X', '8'));
 
     function assertContent(contents) {
-      expect(contents).toMatch(expectedContent);
-    }
-
-    function compareContents(file, enc, cb) {
-      pipe([
-        file.contents,
-        concat(assertContent),
-      ], function(err) {
-        cb(err, file);
-      });
+      expect(contents).toEqual(expectedContent);
     }
 
     function assert(files) {
@@ -211,41 +190,35 @@ describe('.src()', function() {
       expect(files[0].isStream()).toEqual(true);
     }
 
-    pipe([
-      vfs.src(bomInputPath, { removeBOM: false, buffer: false }),
-      through.obj(compareContents),
-      concat(assert),
-    ], done);
+    pipeline(
+      [
+        vfs.src(bomInputPath, { removeBOM: false, buffer: false }),
+        compareContents(assertContent),
+        concatArray(assert),
+      ],
+      done
+    );
   });
 
-  it('removes BOM from utf16be-encoded files by default (buffer)', function(done) {
-    var expectedContent = new Buffer(bomContents.replace('X', '16-BE'));
+  it('removes BOM from utf16be-encoded files by default (buffer)', function (done) {
+    var expectedContent = Buffer.from(bomContents.replace('X', '16-BE'));
 
     function assert(files) {
       expect(files.length).toEqual(1);
-      expect(files[0].contents).toMatch(expectedContent);
+      expect(files[0].contents).toEqual(expectedContent);
     }
 
-    pipe([
-      vfs.src(beBomInputPath, { encoding: 'utf16be' }),
-      concat(assert),
-    ], done);
+    pipeline(
+      [vfs.src(beBomInputPath, { encoding: 'utf16be' }), concatArray(assert)],
+      done
+    );
   });
 
-  it('removes BOM from utf16be-encoded files by default (stream)', function(done) {
-    var expectedContent = new Buffer(bomContents.replace('X', '16-BE'));
+  it('removes BOM from utf16be-encoded files by default (stream)', function (done) {
+    var expectedContent = Buffer.from(bomContents.replace('X', '16-BE'));
 
     function assertContent(contents) {
-      expect(contents).toMatch(expectedContent);
-    }
-
-    function compareContents(file, enc, cb) {
-      pipe([
-        file.contents,
-        concat(assertContent),
-      ], function(err) {
-        cb(err, file);
-      });
+      expect(contents).toEqual(expectedContent);
     }
 
     function assert(files) {
@@ -253,41 +226,42 @@ describe('.src()', function() {
       expect(files[0].isStream()).toEqual(true);
     }
 
-    pipe([
-      vfs.src(beBomInputPath, { encoding: 'utf16be', buffer: false }),
-      through.obj(compareContents),
-      concat(assert),
-    ], done);
+    pipeline(
+      [
+        vfs.src(beBomInputPath, { encoding: 'utf16be', buffer: false }),
+        compareContents(assertContent),
+        concatArray(assert),
+      ],
+      done
+    );
   });
 
-  it('does not remove BOM from utf16be-encoded files if option is false (buffer)', function(done) {
-    var expectedContent = new Buffer('\ufeff' + bomContents.replace('X', '16-BE'));
+  it('does not remove BOM from utf16be-encoded files if option is false (buffer)', function (done) {
+    var expectedContent = Buffer.from(
+      '\ufeff' + bomContents.replace('X', '16-BE')
+    );
 
     function assert(files) {
       expect(files.length).toEqual(1);
-      expect(files[0].contents).toMatch(expectedContent);
+      expect(files[0].contents).toEqual(expectedContent);
     }
 
-    pipe([
-      vfs.src(beBomInputPath, { encoding: 'utf16be', removeBOM: false }),
-      concat(assert),
-    ], done);
+    pipeline(
+      [
+        vfs.src(beBomInputPath, { encoding: 'utf16be', removeBOM: false }),
+        concatArray(assert),
+      ],
+      done
+    );
   });
 
-  it('does not remove BOM from utf16be-encoded files if option is false (stream)', function(done) {
-    var expectedContent = new Buffer('\ufeff' + bomContents.replace('X', '16-BE'));
+  it('does not remove BOM from utf16be-encoded files if option is false (stream)', function (done) {
+    var expectedContent = Buffer.from(
+      '\ufeff' + bomContents.replace('X', '16-BE')
+    );
 
     function assertContent(contents) {
-      expect(contents).toMatch(expectedContent);
-    }
-
-    function compareContents(file, enc, cb) {
-      pipe([
-        file.contents,
-        concat(assertContent),
-      ], function(err) {
-        cb(err, file);
-      });
+      expect(contents).toEqual(expectedContent);
     }
 
     function assert(files) {
@@ -295,41 +269,39 @@ describe('.src()', function() {
       expect(files[0].isStream()).toEqual(true);
     }
 
-    pipe([
-      vfs.src(beBomInputPath, { encoding: 'utf16be', removeBOM: false, buffer: false }),
-      through.obj(compareContents),
-      concat(assert),
-    ], done);
+    pipeline(
+      [
+        vfs.src(beBomInputPath, {
+          encoding: 'utf16be',
+          removeBOM: false,
+          buffer: false,
+        }),
+        compareContents(assertContent),
+        concatArray(assert),
+      ],
+      done
+    );
   });
 
-  it('removes BOM from utf16le-encoded files by default (buffer)', function(done) {
-    var expectedContent = new Buffer(bomContents.replace('X', '16-LE'));
+  it('removes BOM from utf16le-encoded files by default (buffer)', function (done) {
+    var expectedContent = Buffer.from(bomContents.replace('X', '16-LE'));
 
     function assert(files) {
       expect(files.length).toEqual(1);
-      expect(files[0].contents).toMatch(expectedContent);
+      expect(files[0].contents).toEqual(expectedContent);
     }
 
-    pipe([
-      vfs.src(leBomInputPath, { encoding: 'utf16le' }),
-      concat(assert),
-    ], done);
+    pipeline(
+      [vfs.src(leBomInputPath, { encoding: 'utf16le' }), concatArray(assert)],
+      done
+    );
   });
 
-  it('removes BOM from utf16le-encoded files by default (stream)', function(done) {
-    var expectedContent = new Buffer(bomContents.replace('X', '16-LE'));
+  it('removes BOM from utf16le-encoded files by default (stream)', function (done) {
+    var expectedContent = Buffer.from(bomContents.replace('X', '16-LE'));
 
     function assertContent(contents) {
-      expect(contents).toMatch(expectedContent);
-    }
-
-    function compareContents(file, enc, cb) {
-      pipe([
-        file.contents,
-        concat(assertContent),
-      ], function(err) {
-        cb(err, file);
-      });
+      expect(contents).toEqual(expectedContent);
     }
 
     function assert(files) {
@@ -337,41 +309,42 @@ describe('.src()', function() {
       expect(files[0].isStream()).toEqual(true);
     }
 
-    pipe([
-      vfs.src(leBomInputPath, { encoding: 'utf16le', buffer: false }),
-      through.obj(compareContents),
-      concat(assert),
-    ], done);
+    pipeline(
+      [
+        vfs.src(leBomInputPath, { encoding: 'utf16le', buffer: false }),
+        compareContents(assertContent),
+        concatArray(assert),
+      ],
+      done
+    );
   });
 
-  it('does not remove BOM from utf16le-encoded files if option is false (buffer)', function(done) {
-    var expectedContent = new Buffer('\ufeff' + bomContents.replace('X', '16-LE'));
+  it('does not remove BOM from utf16le-encoded files if option is false (buffer)', function (done) {
+    var expectedContent = Buffer.from(
+      '\ufeff' + bomContents.replace('X', '16-LE')
+    );
 
     function assert(files) {
       expect(files.length).toEqual(1);
-      expect(files[0].contents).toMatch(expectedContent);
+      expect(files[0].contents).toEqual(expectedContent);
     }
 
-    pipe([
-      vfs.src(leBomInputPath, { encoding: 'utf16le', removeBOM: false }),
-      concat(assert),
-    ], done);
+    pipeline(
+      [
+        vfs.src(leBomInputPath, { encoding: 'utf16le', removeBOM: false }),
+        concatArray(assert),
+      ],
+      done
+    );
   });
 
-  it('does not remove BOM from utf16le-encoded files if option is false (stream)', function(done) {
-    var expectedContent = new Buffer('\ufeff' + bomContents.replace('X', '16-LE'));
+  it('does not remove BOM from utf16le-encoded files if option is false (stream)', function (done) {
+    var expectedContent = Buffer.from(
+      '\ufeff' + bomContents.replace('X', '16-LE')
+    );
 
     function assertContent(contents) {
-      expect(contents).toMatch(expectedContent);
-    }
-
-    function compareContents(file, enc, cb) {
-      pipe([
-        file.contents,
-        concat(assertContent),
-      ], function(err) {
-        cb(err, file);
-      });
+      expect(contents).toEqual(expectedContent);
     }
 
     function assert(files) {
@@ -379,43 +352,41 @@ describe('.src()', function() {
       expect(files[0].isStream()).toEqual(true);
     }
 
-    pipe([
-      vfs.src(leBomInputPath, { encoding: 'utf16le', removeBOM: false, buffer: false }),
-      through.obj(compareContents),
-      concat(assert),
-    ], done);
+    pipeline(
+      [
+        vfs.src(leBomInputPath, {
+          encoding: 'utf16le',
+          removeBOM: false,
+          buffer: false,
+        }),
+        compareContents(assertContent),
+        concatArray(assert),
+      ],
+      done
+    );
   });
 
   // This goes for any non-UTF-8 encoding.
   // UTF-16-BE is enough to demonstrate this is done properly.
-  it('does not remove anything that looks like a utf8-encoded BOM from utf16be-encoded files (buffer)', function(done) {
+  it('does not remove anything that looks like a utf8-encoded BOM from utf16be-encoded files (buffer)', function (done) {
     var expectedContent = fs.readFileSync(beNotBomInputPath);
 
     function assert(files) {
       expect(files.length).toEqual(1);
-      expect(files[0].contents).toMatch(expectedContent);
-    };
+      expect(files[0].contents).toEqual(expectedContent);
+    }
 
-    pipe([
-      vfs.src(beNotBomInputPath),
-      concat(assert),
-    ], done);
+    pipeline(
+      [vfs.src(beNotBomInputPath, { encoding: false }), concatArray(assert)],
+      done
+    );
   });
 
-  it('does not remove anything that looks like a utf8-encoded BOM from utf16be-encoded files (stream)', function(done) {
+  it('does not remove anything that looks like a utf8-encoded BOM from utf16be-encoded files (stream)', function (done) {
     var expectedContent = fs.readFileSync(beNotBomInputPath);
 
     function assertContent(contents) {
-      expect(contents).toMatch(expectedContent);
-    }
-
-    function compareContents(file, enc, cb) {
-      pipe([
-        file.contents,
-        concat(assertContent),
-      ], function(err) {
-        cb(err, file);
-      });
+      expect(contents).toEqual(expectedContent);
     }
 
     function assert(files) {
@@ -423,43 +394,37 @@ describe('.src()', function() {
       expect(files[0].isStream()).toEqual(true);
     }
 
-    pipe([
-      vfs.src(beNotBomInputPath, { buffer: false }),
-      through.obj(compareContents),
-      concat(assert),
-    ], done);
+    pipeline(
+      [
+        vfs.src(beNotBomInputPath, { buffer: false, encoding: false }),
+        compareContents(assertContent),
+        concatArray(assert),
+      ],
+      done
+    );
   });
 
   // This goes for any non-UTF-8 encoding.
   // UTF-16-LE is enough to demonstrate this is done properly.
-  it('does not remove anything that looks like a utf8-encoded BOM from utf16le-encoded files (buffer)', function(done) {
+  it('does not remove anything that looks like a utf8-encoded BOM from utf16le-encoded files (buffer)', function (done) {
     var expectedContent = fs.readFileSync(leNotBomInputPath);
 
     function assert(files) {
       expect(files.length).toEqual(1);
-      expect(files[0].contents).toMatch(expectedContent);
+      expect(files[0].contents).toEqual(expectedContent);
     }
 
-    pipe([
-      vfs.src(leNotBomInputPath),
-      concat(assert),
-    ], done);
+    pipeline(
+      [vfs.src(leNotBomInputPath, { encoding: false }), concatArray(assert)],
+      done
+    );
   });
 
-  it('does not remove anything that looks like a utf8-encoded BOM from utf16le-encoded files (stream)', function(done) {
+  it('does not remove anything that looks like a utf8-encoded BOM from utf16le-encoded files (stream)', function (done) {
     var expectedContent = fs.readFileSync(leNotBomInputPath);
 
     function assertContent(contents) {
-      expect(contents).toMatch(expectedContent);
-    }
-
-    function compareContents(file, enc, cb) {
-      pipe([
-        file.contents,
-        concat(assertContent),
-      ], function(err) {
-        cb(err, file);
-      });
+      expect(contents).toEqual(expectedContent);
     }
 
     function assert(files) {
@@ -467,41 +432,35 @@ describe('.src()', function() {
       expect(files[0].isStream()).toEqual(true);
     }
 
-    pipe([
-      vfs.src(leNotBomInputPath, { buffer: false }),
-      through.obj(compareContents),
-      concat(assert),
-    ], done);
+    pipeline(
+      [
+        vfs.src(leNotBomInputPath, { buffer: false, encoding: false }),
+        compareContents(assertContent),
+        concatArray(assert),
+      ],
+      done
+    );
   });
 
-  it('does not do any transcoding with encoding option set to false (buffer)', function(done) {
+  it('does not do any transcoding with encoding option set to false (buffer)', function (done) {
     var expectedContents = fs.readFileSync(ranBomInputPath);
 
     function assert(files) {
       expect(files.length).toEqual(1);
-      expect(files[0].contents).toMatch(expectedContents);
+      expect(files[0].contents).toEqual(expectedContents);
     }
 
-    pipe([
-      vfs.src(ranBomInputPath, { encoding: false }),
-      concat(assert),
-    ], done);
+    pipeline(
+      [vfs.src(ranBomInputPath, { encoding: false }), concatArray(assert)],
+      done
+    );
   });
 
-  it('does not do any transcoding with encoding option set to false (stream)', function(done) {
+  it('does not do any transcoding with encoding option set to false (stream)', function (done) {
     var expectedContents = fs.readFileSync(ranBomInputPath);
 
     function assertContent(contents) {
-      expect(contents).toMatch(expectedContents);
-    }
-
-    function compareContents(file, enc, cb) {
-      pipe([
-        file.contents,
-        concat(assertContent),
-      ], function(err) {
-        cb(err, file);
-      });
+      expect(contents).toEqual(expectedContents);
     }
 
     function assert(files) {
@@ -509,83 +468,70 @@ describe('.src()', function() {
       expect(files[0].isStream()).toEqual(true);
     }
 
-    pipe([
-      vfs.src(ranBomInputPath, { encoding: false, buffer: false }),
-      through.obj(compareContents),
-      concat(assert),
-    ], done);
+    pipeline(
+      [
+        vfs.src(ranBomInputPath, { encoding: false, buffer: false }),
+        compareContents(assertContent),
+        concatArray(assert),
+      ],
+      done
+    );
   });
 
-  it('does not remove utf8 BOM with encoding option set to false (buffer)', function(done) {
+  it('does not remove utf8 BOM with encoding option set to false (buffer)', function (done) {
     var expectedContents = fs.readFileSync(bomInputPath);
 
     function assert(files) {
       expect(files.length).toEqual(1);
-      expect(files[0].contents).toMatch(expectedContents);
+      expect(files[0].contents).toEqual(expectedContents);
     }
 
-    pipe([
-      vfs.src(bomInputPath, { encoding: false }),
-      concat(assert),
-    ], done);
+    pipeline(
+      [vfs.src(bomInputPath, { encoding: false }), concatArray(assert)],
+      done
+    );
   });
 
-  it('does not remove utf8 BOM with encoding option set to false (stream)', function(done) {
+  it('does not remove utf8 BOM with encoding option set to false (stream)', function (done) {
     var expectedContents = fs.readFileSync(bomInputPath);
 
     function assertContent(contents) {
-      expect(contents).toMatch(expectedContents);
+      expect(contents).toEqual(expectedContents);
     }
-
-    function compareContents(file, enc, cb) {
-      pipe([
-        file.contents,
-        concat(assertContent),
-      ], function(err) {
-        cb(err, file);
-      });
-    }
-
     function assert(files) {
       expect(files.length).toEqual(1);
       expect(files[0].isStream()).toEqual(true);
     }
 
-    pipe([
-      vfs.src(bomInputPath, { encoding: false, buffer: false }),
-      through.obj(compareContents),
-      concat(assert),
-    ], done);
+    pipeline(
+      [
+        vfs.src(bomInputPath, { encoding: false, buffer: false }),
+        compareContents(assertContent),
+        concatArray(assert),
+      ],
+      done
+    );
   });
 
-  it('transcodes gb2312 to utf8 with encoding option (buffer)', function(done) {
-    var expectedContents = new Buffer(encodedContents);
+  it('transcodes gb2312 to utf8 with encoding option (buffer)', function (done) {
+    var expectedContents = Buffer.from(encodedContents);
 
     function assert(files) {
       expect(files.length).toEqual(1);
-      expect(files[0].contents).toMatch(expectedContents);
+      expect(files[0].contents).toEqual(expectedContents);
     }
 
-    pipe([
-      vfs.src(encodedInputPath, { encoding: 'gb2312' }),
-      concat(assert),
-    ], done);
+    pipeline(
+      [vfs.src(encodedInputPath, { encoding: 'gb2312' }), concatArray(assert)],
+      done
+    );
   });
 
-  it('transcodes gb2312 to utf8 with encoding option (stream)', function(done) {
-    var expectedContents = new Buffer(encodedContents);
+  it('transcodes gb2312 to utf8 with encoding option (stream)', function (done) {
+    var expectedContents = Buffer.from(encodedContents);
 
     function assertContent(contents) {
-      expect(contents).toMatch(expectedContents);
-    }
-
-    function compareContents(file, enc, cb) {
-      pipe([
-        file.contents,
-        concat(assertContent),
-      ], function(err) {
-        cb(err, file);
-      });
+      expect(contents).toEqual(expectedContents);
     }
 
     function assert(files) {
@@ -593,74 +539,79 @@ describe('.src()', function() {
       expect(files[0].isStream()).toEqual(true);
     }
 
-    pipe([
-      vfs.src(encodedInputPath, { encoding: 'gb2312', buffer: false }),
-      through.obj(compareContents),
-      concat(assert),
-    ], done);
+    pipeline(
+      [
+        vfs.src(encodedInputPath, { encoding: 'gb2312', buffer: false }),
+        compareContents(assertContent),
+        concatArray(assert),
+      ],
+      done
+    );
   });
 
-  it('reports unsupported encoding errors (buffer)', function(done) {
+  it('reports unsupported encoding errors (buffer)', function (done) {
     function assert(files) {
       expect(files.length).toEqual(0);
     }
 
     function finish(err) {
-      expect(err).toExist();
+      expect(err).toEqual(expect.anything());
       expect(err.message).toEqual('Unsupported encoding: fubar42');
       done();
     }
 
-    pipe([
-      vfs.src(inputPath, { encoding: 'fubar42' }),
-      concat(assert),
-    ], finish);
+    pipeline(
+      [vfs.src(inputPath, { encoding: 'fubar42' }), concatArray(assert)],
+      finish
+    );
   });
 
-  it('reports unsupported encoding errors (stream)', function(done) {
+  it('reports unsupported encoding errors (stream)', function (done) {
     function assert(files) {
       expect(files.length).toEqual(0);
     }
 
     function finish(err) {
-      expect(err).toExist();
+      expect(err).toEqual(expect.anything());
       expect(err.message).toEqual('Unsupported encoding: fubar42');
       done();
     }
 
-    pipe([
-      vfs.src(inputPath, { encoding: 'fubar42', buffer: false }),
-      concat(assert),
-    ], finish);
+    pipeline(
+      [
+        vfs.src(inputPath, { encoding: 'fubar42', buffer: false }),
+        concatArray(assert),
+      ],
+      finish
+    );
   });
 
-
-  it('globs files with default settings', function(done) {
+  it('globs files with default settings', function (done) {
     function assert(files) {
       expect(files.length).toEqual(7);
     }
 
-    pipe([
-      vfs.src('./fixtures/*.txt', { cwd: __dirname }),
-      concat(assert),
-    ], done);
+    pipeline(
+      [vfs.src('./fixtures/*.txt', { cwd: __dirname }), concatArray(assert)],
+      done
+    );
   });
 
-  it('globs files with default settings and relative cwd', function(done) {
+  it('globs files with default settings and relative cwd', function (done) {
     var cwd = path.relative(process.cwd(), __dirname);
 
     function assert(files) {
       expect(files.length).toEqual(7);
     }
 
-    pipe([
-      vfs.src('./fixtures/*.txt', { cwd: cwd }),
-      concat(assert),
-    ], done);
+    pipeline(
+      [vfs.src('./fixtures/*.txt', { cwd: cwd }), concatArray(assert)],
+      done
+    );
   });
 
   // TODO: need to normalize the path of a directory vinyl object
-  it('globs a directory with default settings', function(done) {
+  it('globs a directory with default settings', function (done) {
     var inputDirGlob = path.join(inputBase, './f*/');
 
     function assert(files) {
@@ -669,13 +620,10 @@ describe('.src()', function() {
       expect(files[0].isDirectory()).toEqual(true);
     }
 
-    pipe([
-      vfs.src(inputDirGlob),
-      concat(assert),
-    ], done);
+    pipeline([vfs.src(inputDirGlob), concatArray(assert)], done);
   });
 
-  it('globs a directory with default settings and relative cwd', function(done) {
+  it('globs a directory with default settings and relative cwd', function (done) {
     var cwd = path.relative(process.cwd(), __dirname);
 
     function assert(files) {
@@ -684,13 +632,13 @@ describe('.src()', function() {
       expect(files[0].isDirectory()).toEqual(true);
     }
 
-    pipe([
-      vfs.src('./fixtures/f*/', { cwd: cwd }),
-      concat(assert),
-    ], done);
+    pipeline(
+      [vfs.src('./fixtures/f*/', { cwd: cwd }), concatArray(assert)],
+      done
+    );
   });
 
-  it('streams a directory with default settings', function(done) {
+  it('streams a directory with default settings', function (done) {
     function assert(files) {
       expect(files.length).toEqual(1);
       expect(files[0].path).toEqual(inputDirpath);
@@ -698,27 +646,21 @@ describe('.src()', function() {
       expect(files[0].isDirectory()).toEqual(true);
     }
 
-    pipe([
-      vfs.src(inputDirpath),
-      concat(assert),
-    ], done);
+    pipeline([vfs.src(inputDirpath), concatArray(assert)], done);
   });
 
-  it('streams file with with no contents using read: false option', function(done) {
+  it('streams file with with no contents using read: false option', function (done) {
     function assert(files) {
       expect(files.length).toEqual(1);
       expect(files[0].path).toEqual(inputPath);
       expect(files[0].isNull()).toEqual(true);
-      expect(files[0].contents).toNotExist();
+      expect(files[0].contents).toEqual(null);
     }
 
-    pipe([
-      vfs.src(inputPath, { read: false }),
-      concat(assert),
-    ], done);
+    pipeline([vfs.src(inputPath, { read: false }), concatArray(assert)], done);
   });
 
-  it('streams a file changed after since', function(done) {
+  it('streams a file changed after since', function (done) {
     var lastUpdateDate = new Date(+fs.statSync(inputPath).mtime - 1000);
 
     function assert(files) {
@@ -726,39 +668,30 @@ describe('.src()', function() {
       expect(files[0].path).toEqual(inputPath);
     }
 
-    pipe([
-      vfs.src(inputPath, { since: lastUpdateDate }),
-      concat(assert),
-    ], done);
+    pipeline(
+      [vfs.src(inputPath, { since: lastUpdateDate }), concatArray(assert)],
+      done
+    );
   });
 
-  it('does not stream a file changed before since', function(done) {
+  it('does not stream a file changed before since', function (done) {
     var lastUpdateDate = new Date(+fs.statSync(inputPath).mtime + 1000);
 
     function assert(files) {
       expect(files.length).toEqual(0);
     }
 
-    pipe([
-      vfs.src(inputPath, { since: lastUpdateDate }),
-      concat(assert),
-    ], done);
+    pipeline(
+      [vfs.src(inputPath, { since: lastUpdateDate }), concatArray(assert)],
+      done
+    );
   });
 
-  it('streams a file with streaming contents', function(done) {
+  it('streams a file with streaming contents', function (done) {
     var expectedContent = fs.readFileSync(inputPath);
 
     function assertContent(contents) {
-      expect(contents).toMatch(expectedContent);
-    }
-
-    function compareContents(file, enc, cb) {
-      pipe([
-        file.contents,
-        concat(assertContent),
-      ], function(err) {
-        cb(err, file);
-      });
+      expect(contents).toEqual(expectedContent);
     }
 
     function assert(files) {
@@ -767,14 +700,17 @@ describe('.src()', function() {
       expect(files[0].isStream()).toEqual(true);
     }
 
-    pipe([
-      vfs.src(inputPath, { buffer: false }),
-      through.obj(compareContents),
-      concat(assert),
-    ], done);
+    pipeline(
+      [
+        vfs.src(inputPath, { buffer: false }),
+        compareContents(assertContent),
+        concatArray(assert),
+      ],
+      done
+    );
   });
 
-  it('can be used as a through stream and adds new files to the end', function(done) {
+  it('can be used as a through stream and adds new files to the end', function (done) {
     var file = new File({
       base: inputBase,
       path: inputPath,
@@ -787,37 +723,29 @@ describe('.src()', function() {
       expect(files[0]).toEqual(file);
     }
 
-    pipe([
-      from.obj([file]),
-      vfs.src(inputPath),
-      concat(assert),
-    ], done);
+    pipeline([from([file]), vfs.src(inputPath), concatArray(assert)], done);
   });
 
-  it('can be used at beginning and in the middle', function(done) {
+  it('can be used at beginning and in the middle', function (done) {
     function assert(files) {
       expect(files.length).toEqual(2);
     }
 
-    pipe([
-      vfs.src(inputPath),
-      vfs.src(inputPath),
-      concat(assert),
-    ], done);
+    pipeline(
+      [vfs.src(inputPath), vfs.src(inputPath), concatArray(assert)],
+      done
+    );
   });
 
-  it('does not pass options on to through2', function(done) {
+  it('does not pass options on to stream.Transform', function (done) {
     // Reference: https://github.com/gulpjs/vinyl-fs/issues/153
-    var read = expect.createSpy().andReturn(false);
+    var read = sinon.fake.returns(false);
 
     function assert() {
       // Called once to resolve the option
-      expect(read.calls.length).toEqual(1);
+      expect(read.callCount).toEqual(1);
     }
 
-    pipe([
-      vfs.src(inputPath, { read: read }),
-      concat(assert),
-    ], done);
+    pipeline([vfs.src(inputPath, { read: read }), concatArray(assert)], done);
   });
 });
