@@ -3,7 +3,9 @@
 /* eslint-disable new-cap */
 
 var iconv = require('iconv-lite');
-var through = require('through2');
+// TODO: Reaching into iconv-lite is fragile
+var StripBOM = require('iconv-lite/lib/bom-handling').StripBOM;
+var Transform = require('streamx').Transform;
 
 var DEFAULT_ENCODING = require('./constants').DEFAULT_ENCODING;
 
@@ -13,78 +15,79 @@ function Codec(codec, encoding) {
   this.bomAware = codec.bomAware || false;
 }
 
-
 function getEncoder(codec) {
   return new codec.encoder(null, codec);
 }
 
-Codec.prototype.encode = function(str) {
+Codec.prototype.encode = function (str) {
   var encoder = getEncoder(this.codec);
   var buf = encoder.write(str);
   var end = encoder.end();
   return end && end.length > 0 ? Buffer.concat(buf, end) : buf;
 };
 
-Codec.prototype.encodeStream = function() {
+Codec.prototype.encodeStream = function () {
   var encoder = getEncoder(this.codec);
-  return through(
-    { decodeStrings: false },
-    function(str, enc, cb) {
+  return new Transform({
+    transform: function (str, cb) {
       var buf = encoder.write(str);
       if (buf && buf.length) {
         this.push(buf);
       }
       cb();
     },
-    function(cb) {
+    flush: function (cb) {
       var buf = encoder.end();
       if (buf && buf.length) {
         this.push(buf);
       }
       cb();
-    }
-  );
+    },
+  });
 };
-
 
 function getDecoder(codec) {
   return new codec.decoder(null, codec);
 }
 
-Codec.prototype.decode = function(buf) {
+Codec.prototype.decode = function (buf, options) {
   var decoder = getDecoder(this.codec);
+  if (options && options.removeBOM) {
+    decoder = new StripBOM(decoder);
+  }
   var str = decoder.write(buf);
   var end = decoder.end();
   return end ? str + end : str;
 };
 
-Codec.prototype.decodeStream = function() {
+Codec.prototype.decodeStream = function (options) {
   var decoder = getDecoder(this.codec);
-  return through(
-    { encoding: DEFAULT_ENCODING },
-    function(buf, enc, cb) {
+  if (options && options.removeBOM) {
+    decoder = new StripBOM(decoder);
+  }
+  return new Transform({
+    transform: function (buf, cb) {
       var str = decoder.write(buf);
       if (str && str.length) {
-        this.push(str, DEFAULT_ENCODING);
+        this.push(str);
       }
       cb();
     },
-    function(cb) {
+    flush: function (cb) {
       var str = decoder.end();
       if (str && str.length) {
-        this.push(str, DEFAULT_ENCODING);
+        this.push(str);
       }
       cb();
-    }
-  );
+    },
+  });
 };
-
 
 var cache = {};
 
 function getCodec(encoding) {
   var codec = cache[encoding];
-  if (!!codec || !encoding || cache.hasOwnProperty(encoding)) {
+  if (!!codec || !encoding || Object.hasOwnProperty.call(cache, encoding)) {
     return codec;
   }
 
@@ -97,7 +100,6 @@ function getCodec(encoding) {
   cache[encoding] = codec;
   return codec;
 }
-
 
 // Pre-load default encoding
 getCodec(DEFAULT_ENCODING);
