@@ -1,7 +1,7 @@
 'use strict';
 
 var lead = require('lead');
-var pumpify = require('pumpify');
+var composer = require('stream-composer');
 var mkdirpStream = require('fs-mkdirp-stream');
 var createResolver = require('resolve-options');
 
@@ -31,10 +31,10 @@ function dest(outFolder, opt) {
     callback(null, file.dirname, dirMode);
   }
 
-  var saveStream = pumpify.obj(
+  var saveStream = composer.pipeline(
     prepare(folderResolver, optResolver),
     sourcemap(optResolver),
-    mkdirpStream.obj(dirpath),
+    mkdirpStream(dirpath),
     writeContents(optResolver)
   );
 
