@@ -4,14 +4,14 @@ var path = require('path');
 
 var fs = require('graceful-fs');
 var Vinyl = require('vinyl');
-var through = require('through2');
+var Transform = require('streamx').Transform;
 
 function prepareWrite(folderResolver, optResolver) {
   if (!folderResolver) {
     throw new Error('Invalid output folder');
   }
 
-  function normalize(file, enc, cb) {
+  function normalize(file, cb) {
     if (!Vinyl.isVinyl(file)) {
       return cb(new Error('Received a non-Vinyl object in `dest()`'));
     }
@@ -42,7 +42,9 @@ function prepareWrite(folderResolver, optResolver) {
     cb(null, file);
   }
 
-  return through.obj(normalize);
+  return new Transform({
+    transform: normalize,
+  });
 }
 
 module.exports = prepareWrite;
