@@ -1,11 +1,11 @@
 'use strict';
 
-var through = require('through2');
+var Transform = require('streamx').Transform;
 var sourcemap = require('vinyl-sourcemap');
 
 function sourcemapStream(optResolver) {
 
-  function saveSourcemap(file, enc, callback) {
+  function saveSourcemap(file, callback) {
     var self = this;
 
     var srcMap = optResolver.resolve('sourcemaps', file);
@@ -32,7 +32,9 @@ function sourcemapStream(optResolver) {
     }
   }
 
-  return through.obj(saveSourcemap);
+  return new Transform({
+    transform: saveSourcemap,
+  });
 }
 
 module.exports = sourcemapStream;
