@@ -1,6 +1,6 @@
 'use strict';
 
-var through = require('through2');
+var Transform = require('streamx').Transform;
 
 var writeDir = require('./write-dir');
 var writeStream = require('./write-stream');
@@ -11,7 +11,7 @@ var fo = require('../../file-operations');
 
 function writeContents(optResolver) {
 
-  function writeFile(file, enc, callback) {
+  function writeFile(file, callback) {
     // Write it as a symlink
     if (file.isSymbolic()) {
       return writeSymbolicLink(file, optResolver, onWritten);
@@ -53,7 +53,9 @@ function writeContents(optResolver) {
 
   }
 
-  return through.obj(writeFile);
+  return new Transform({
+    transform: writeFile,
+  });
 }
 
 module.exports = writeContents;
