@@ -1,5 +1,7 @@
 'use strict';
 
+var pipeline = require('streamx').pipeline;
+
 var fo = require('../../file-operations');
 var getCodec = require('../../codecs');
 var DEFAULT_ENCODING = require('../../constants').DEFAULT_ENCODING;
@@ -26,41 +28,20 @@ function writeStream(file, optResolver, onWritten) {
   // TODO: is this the best API?
   var outStream = fo.createWriteStream(file.path, opt, onFlush);
 
-  var contents = file.contents;
+  // TODO: should this use a clone?
+  var streams = [file.contents];
 
   if (encoding && encoding.enc !== DEFAULT_ENCODING) {
-    contents = contents
-      .pipe(getCodec(DEFAULT_ENCODING).decodeStream())
-      .pipe(codec.encodeStream());
+    streams.push(getCodec(DEFAULT_ENCODING).decodeStream());
+    streams.push(codec.encodeStream());
   }
 
-  file.contents.once('error', onComplete);
-  outStream.once('error', onComplete);
-  outStream.once('finish', onComplete);
+  streams.push(outStream);
 
-  // TODO: should this use a clone?
-  contents.pipe(outStream);
-
-  function onComplete(streamErr) {
-    // Cleanup event handlers before closing
-    file.contents.removeListener('error', onComplete);
-    outStream.removeListener('error', onComplete);
-    outStream.removeListener('finish', onComplete);
-
-    // Need to guarantee the fd is closed before forwarding the error
-    outStream.once('close', onClose);
-    outStream.end();
-
-    function onClose(closeErr) {
-      onWritten(streamErr || closeErr);
-    }
-  }
+  pipeline(streams, onWritten);
 
   // Cleanup
   function onFlush(fd, callback) {
-    // TODO: removing this before readStream because it replaces the stream
-    file.contents.removeListener('error', onComplete);
-
     // TODO: this is doing sync stuff & the callback seems unnecessary
     readStream(file, { resolve: resolve }, complete);
 
@@ -71,7 +52,7 @@ function writeStream(file, optResolver, onWritten) {
       if (key === 'removeBOM') {
         return false;
       }
-      throw new Error('Eek! stub resolver doesn\'t have ' + key);
+      throw new Error("Eek! stub resolver doesn't have " + key);
     }
 
     function complete() {
@@ -82,7 +63,6 @@ function writeStream(file, optResolver, onWritten) {
       fo.updateMetadata(fd, file, callback);
     }
   }
-
 }
 
 module.exports = writeStream;
