@@ -1,11 +1,8 @@
 'use strict';
 
-var util = require('util');
-
 var fs = require('graceful-fs');
-var assign = require('object.assign');
 var date = require('value-or-function').date;
-var Writable = require('readable-stream').Writable;
+var Writable = require('streamx').Writable;
 
 var constants = require('./constants');
 
@@ -80,15 +77,13 @@ function getModeDiff(fsMode, vinylMode) {
 }
 
 function getTimesDiff(fsStat, vinylStat) {
-
   var mtime = date(vinylStat.mtime) || 0;
   if (!mtime) {
     return;
   }
 
   var atime = date(vinylStat.atime) || 0;
-  if (+mtime === +fsStat.mtime &&
-      +atime === +fsStat.atime) {
+  if (+mtime === +fsStat.mtime && +atime === +fsStat.atime) {
     return;
   }
 
@@ -105,13 +100,14 @@ function getTimesDiff(fsStat, vinylStat) {
 }
 
 function getOwnerDiff(fsStat, vinylStat) {
-  if (!isValidUnixId(vinylStat.uid) &&
-      !isValidUnixId(vinylStat.gid)) {
+  if (!isValidUnixId(vinylStat.uid) && !isValidUnixId(vinylStat.gid)) {
     return;
   }
 
-  if ((!isValidUnixId(fsStat.uid) && !isValidUnixId(vinylStat.uid)) ||
-      (!isValidUnixId(fsStat.gid) && !isValidUnixId(vinylStat.gid))) {
+  if (
+    (!isValidUnixId(fsStat.uid) && !isValidUnixId(vinylStat.uid)) ||
+    (!isValidUnixId(fsStat.gid) && !isValidUnixId(vinylStat.gid))
+  ) {
     return;
   }
 
@@ -125,8 +121,7 @@ function getOwnerDiff(fsStat, vinylStat) {
     gid = vinylStat.gid;
   }
 
-  if (uid === fsStat.uid &&
-      gid === fsStat.gid) {
+  if (uid === fsStat.uid && gid === fsStat.gid) {
     return;
   }
 
@@ -139,8 +134,8 @@ function getOwnerDiff(fsStat, vinylStat) {
 }
 
 function isOwner(fsStat) {
-  var hasGetuid = (typeof process.getuid === 'function');
-  var hasGeteuid = (typeof process.geteuid === 'function');
+  var hasGetuid = typeof process.getuid === 'function';
+  var hasGeteuid = typeof process.geteuid === 'function';
 
   // If we don't have either, assume we don't have permissions.
   // This should only happen on Windows.
@@ -161,6 +156,32 @@ function isOwner(fsStat) {
   }
 
   return true;
+}
+
+// Node 10 on Windows fails with EPERM if you stat a symlink to a directory so we recursively readlink before we reflect stats
+// TODO: Remove this indirection when we drop Node 10 support
+function findSymlinkHardpath(path, callback) {
+  fs.readlink(path, onReadlink);
+
+  function onReadlink(readlinkErr, resolvedPath) {
+    if (readlinkErr) {
+      return callback(readlinkErr);
+    }
+
+    fs.lstat(resolvedPath, onLstat);
+
+    function onLstat(lstatErr, stat) {
+      if (lstatErr) {
+        return callback(lstatErr);
+      }
+
+      if (stat.isSymbolicLink()) {
+        return findSymlinkHardpath(resolvedPath, callback);
+      }
+
+      callback(null, resolvedPath);
+    }
+  }
 }
 
 function reflectStat(path, file, callback) {
@@ -192,7 +213,6 @@ function reflectLinkStat(path, file, callback) {
 }
 
 function updateMetadata(fd, file, callback) {
-
   fs.fstat(fd, onStat);
 
   function onStat(statErr, stat) {
@@ -210,7 +230,7 @@ function updateMetadata(fd, file, callback) {
     var ownerDiff = getOwnerDiff(stat, file.stat);
 
     // Set file.stat to the reflect current state on disk
-    assign(file.stat, stat);
+    Object.assign(file.stat, stat);
 
     // Nothing to do
     if (!modeDiff && !timesDiff && !ownerDiff) {
@@ -343,139 +363,74 @@ function writeFile(filepath, data, options, callback) {
   }
 }
 
-function createWriteStream(path, options, flush) {
-  return new WriteStream(path, options, flush);
+function noopFlush(fd, cb) {
+  cb();
 }
 
-// Taken from node core and altered to receive a flush function and simplified
-// To be used for cleanup (like updating times/mode/etc)
-function WriteStream(path, options, flush) {
-  // Not exposed so we can avoid the case where someone doesn't use `new`
-
+function createWriteStream(path, options, flush) {
   if (typeof options === 'function') {
     flush = options;
     options = null;
   }
 
   options = options || {};
+  flush = flush || noopFlush;
 
-  Writable.call(this, options);
+  var mode = options.mode || constants.DEFAULT_FILE_MODE;
+  var flags = options.flags || 'w';
 
-  this.flush = flush;
-  this.path = path;
+  var fd = null;
 
-  this.mode = options.mode || constants.DEFAULT_FILE_MODE;
-  this.flags = options.flags || 'w';
+  return new Writable({
+    mapWritable: function (data) {
+      if (typeof data === 'string') {
+        return Buffer.from(data);
+      } else {
+        return data;
+      }
+    },
+    open: function (cb) {
+      fs.open(path, flags, mode, onOpen);
 
-  // Used by node's `fs.WriteStream`
-  this.fd = null;
-  this.start = null;
+      function onOpen(openErr, openedFd) {
+        if (openErr) {
+          cb(openErr);
+          return;
+        }
 
-  this.open();
+        fd = openedFd;
+        cb();
+      }
+    },
+    destroy: function (cb) {
+      if (fd) {
+        fs.close(fd, onClose);
+      } else {
+        onClose();
+      }
 
-  // Dispose on finish.
-  this.once('finish', this.close);
-}
+      function onClose(closeErr) {
+        fd = null;
+        cb(closeErr);
+      }
+    },
+    write: function (data, cb) {
+      fs.write(fd, data, 0, data.length, null, onWrite);
 
-util.inherits(WriteStream, Writable);
+      function onWrite(writeErr) {
+        if (writeErr) {
+          cb(writeErr);
+          return;
+        }
 
-WriteStream.prototype.open = function() {
-  var self = this;
-
-  fs.open(this.path, this.flags, this.mode, onOpen);
-
-  function onOpen(openErr, fd) {
-    if (openErr) {
-      self.destroy();
-      self.emit('error', openErr);
-      return;
-    }
-
-    self.fd = fd;
-    self.emit('open', fd);
-  }
-};
-
-// Use our `end` method since it is patched for flush
-WriteStream.prototype.destroySoon = WriteStream.prototype.end;
-
-WriteStream.prototype._destroy = function(err, cb) {
-  this.close(function(err2) {
-    cb(err || err2);
+        cb();
+      }
+    },
+    final: function (cb) {
+      flush(fd, cb);
+    },
   });
-};
-
-WriteStream.prototype.close = function(cb) {
-  var that = this;
-
-  if (cb) {
-    this.once('close', cb);
-  }
-
-  if (this.closed || typeof this.fd !== 'number') {
-    if (typeof this.fd !== 'number') {
-      this.once('open', closeOnOpen);
-      return;
-    }
-
-    return process.nextTick(function() {
-      that.emit('close');
-    });
-  }
-
-  this.closed = true;
-
-  fs.close(this.fd, function(er) {
-    if (er) {
-      that.emit('error', er);
-    } else {
-      that.emit('close');
-    }
-  });
-
-  this.fd = null;
-};
-
-WriteStream.prototype._final = function(callback) {
-  if (typeof this.flush !== 'function') {
-    return callback();
-  }
-
-  this.flush(this.fd, callback);
-};
-
-function closeOnOpen() {
-  this.close();
 }
-
-WriteStream.prototype._write = function(data, encoding, callback) {
-  var self = this;
-
-  // This is from node core but I have no idea how to get code coverage on it
-  if (!Buffer.isBuffer(data)) {
-    return this.emit('error', new Error('Invalid data'));
-  }
-
-  if (typeof this.fd !== 'number') {
-    return this.once('open', onOpen);
-  }
-
-  fs.write(this.fd, data, 0, data.length, null, onWrite);
-
-  function onOpen() {
-    self._write(data, encoding, callback);
-  }
-
-  function onWrite(writeErr) {
-    if (writeErr) {
-      self.destroy();
-      callback(writeErr);
-      return;
-    }
-
-    callback();
-  }
-};
 
 module.exports = {
   closeFd: closeFd,
@@ -487,6 +442,7 @@ module.exports = {
   getTimesDiff: getTimesDiff,
   getOwnerDiff: getOwnerDiff,
   isOwner: isOwner,
+  findSymlinkHardpath: findSymlinkHardpath,
   reflectStat: reflectStat,
   reflectLinkStat: reflectLinkStat,
   updateMetadata: updateMetadata,
