@@ -1,9 +1,10 @@
 'use strict';
 
 var gs = require('glob-stream');
-var pumpify = require('pumpify');
+var pipeline = require('streamx').pipeline;
 var toThrough = require('to-through');
 var isValidGlob = require('is-valid-glob');
+var normalizePath = require('normalize-path');
 var createResolver = require('resolve-options');
 
 var config = require('./options');
@@ -20,19 +21,18 @@ function src(glob, opt) {
     throw new Error('Invalid glob argument: ' + glob);
   }
 
-  var streams = [
+  glob = normalizePath(glob, false);
+
+  var outputStream = pipeline(
     gs(glob, opt),
     wrapVinyl(optResolver),
     resolveSymlinks(optResolver),
     prepare(optResolver),
     readContents(optResolver),
-    sourcemap(optResolver),
-  ];
-
-  var outputStream = pumpify.obj(streams);
+    sourcemap(optResolver)
+  );
 
   return toThrough(outputStream);
 }
-
 
 module.exports = src;
