@@ -1,10 +1,10 @@
 'use strict';
 
-var through = require('through2');
+var Transform = require('streamx').Transform;
 
 function prepareRead(optResolver) {
 
-  function normalize(file, enc, callback) {
+  function normalize(file, callback) {
 
     var since = optResolver.resolve('since', file);
 
@@ -16,7 +16,9 @@ function prepareRead(optResolver) {
     return callback(null, file);
   }
 
-  return through.obj(normalize);
+  return new Transform({
+    transform: normalize,
+  });
 }
 
 module.exports = prepareRead;
