@@ -1,6 +1,6 @@
 'use strict';
 
-var through = require('through2');
+var Transform = require('streamx').Transform;
 
 var readDir = require('./read-dir');
 var readStream = require('./read-stream');
@@ -9,7 +9,7 @@ var readSymbolicLink = require('./read-symbolic-link');
 
 function readContents(optResolver) {
 
-  function readFile(file, enc, callback) {
+  function readFile(file, callback) {
 
     // Skip reading contents if read option says so
     var read = optResolver.resolve('read', file);
@@ -46,7 +46,9 @@ function readContents(optResolver) {
     }
   }
 
-  return through.obj(readFile);
+  return new Transform({
+    transform: readFile,
+  });
 }
 
 module.exports = readContents;
