@@ -1,7 +1,6 @@
 'use strict';
 
 var fs = require('graceful-fs');
-var removeBomBuffer = require('remove-bom-buffer');
 
 var getCodec = require('../../codecs');
 var DEFAULT_ENCODING = require('../../constants').DEFAULT_ENCODING;
@@ -23,14 +22,9 @@ function bufferFile(file, optResolver, onRead) {
     if (encoding) {
       var removeBOM = codec.bomAware && optResolver.resolve('removeBOM', file);
 
-      if (codec.enc !== DEFAULT_ENCODING) {
-        contents = codec.decode(contents);
-        removeBOM = removeBOM && contents[0] === '\ufeff';
+      if (removeBOM || codec.enc !== DEFAULT_ENCODING) {
+        contents = codec.decode(contents, { removeBOM: removeBOM });
         contents = getCodec(DEFAULT_ENCODING).encode(contents);
-      }
-
-      if (removeBOM) {
-        contents = removeBomBuffer(contents);
       }
     }
 
