@@ -1,7 +1,7 @@
 'use strict';
 
 var fs = require('graceful-fs');
-var removeBomStream = require('remove-bom-stream');
+var composer = require('stream-composer');
 var lazystream = require('lazystream');
 
 var getCodec = require('../../codecs');
@@ -16,24 +16,25 @@ function streamFile(file, optResolver, onRead) {
 
   var filePath = file.path;
 
-  file.contents = new lazystream.Readable(function() {
+  file.contents = new lazystream.Readable(function () {
     var contents = fs.createReadStream(filePath);
+
+    var streams = [contents];
 
     if (encoding) {
       var removeBOM = codec.bomAware && optResolver.resolve('removeBOM', file);
 
-      if (codec.enc !== DEFAULT_ENCODING) {
-        contents = contents
-          .pipe(codec.decodeStream())
-          .pipe(getCodec(DEFAULT_ENCODING).encodeStream());
-      }
-
-      if (removeBOM) {
-        contents = contents.pipe(removeBomStream());
+      if (removeBOM || codec.enc !== DEFAULT_ENCODING) {
+        streams.push(codec.decodeStream({ removeBOM: removeBOM }));
+        streams.push(getCodec(DEFAULT_ENCODING).encodeStream());
       }
     }
 
-    return contents;
+    if (streams.length > 1) {
+      return composer.pipeline(streams);
+    } else {
+      return contents;
+    }
   });
 
   onRead();
