@@ -1,12 +1,12 @@
 'use strict';
 
-var through = require('through2');
+var Transform = require('streamx').Transform;
 var fo = require('../file-operations');
 
 function resolveSymlinks(optResolver) {
 
   // A stat property is exposed on file objects as a (wanted) side effect
-  function resolveFile(file, enc, callback) {
+  function resolveFile(file, callback) {
 
     fo.reflectLinkStat(file.path, file, onReflect);
 
@@ -25,12 +25,22 @@ function resolveSymlinks(optResolver) {
         return callback(null, file);
       }
 
+      fo.findSymlinkHardpath(file.path, onSymlinkHardpath);
+    }
+
+    function onSymlinkHardpath(readlinkErr, path) {
+      if (readlinkErr) {
+        return callback(readlinkErr);
+      }
+
       // Get target's stats
-      fo.reflectStat(file.path, file, onReflect);
+      fo.reflectStat(path, file, onReflect);
     }
   }
 
-  return through.obj(resolveFile);
+  return new Transform({
+    transform: resolveFile,
+  });
 }
 
 module.exports = resolveSymlinks;
