@@ -1,11 +1,11 @@
 'use strict';
 
-var through = require('through2');
+var Transform = require('streamx').Transform;
 var sourcemap = require('vinyl-sourcemap');
 
 function sourcemapStream(optResolver) {
 
-  function addSourcemap(file, enc, callback) {
+  function addSourcemap(file, callback) {
     var srcMap = optResolver.resolve('sourcemaps', file);
 
     if (!srcMap) {
@@ -23,7 +23,9 @@ function sourcemapStream(optResolver) {
     }
   }
 
-  return through.obj(addSourcemap);
+  return new Transform({
+    transform: addSourcemap,
+  });
 }
 
 module.exports = sourcemapStream;
