@@ -1,18 +1,20 @@
 'use strict';
 
 var File = require('vinyl');
-var through = require('through2');
+var Transform = require('streamx').Transform;
 
 function wrapVinyl() {
 
-  function wrapFile(globFile, enc, callback) {
+  function wrapFile(globFile, callback) {
 
     var file = new File(globFile);
 
     callback(null, file);
   }
 
-  return through.obj(wrapFile);
+  return new Transform({
+    transform: wrapFile,
+  });
 }
 
 module.exports = wrapVinyl;
