@@ -1,7 +1,7 @@
 'use strict';
 
-var pumpify = require('pumpify');
 var lead = require('lead');
+var composer = require('stream-composer');
 var mkdirpStream = require('fs-mkdirp-stream');
 var createResolver = require('resolve-options');
 
@@ -30,9 +30,9 @@ function symlink(outFolder, opt) {
     callback(null, file.dirname, dirMode);
   }
 
-  var stream = pumpify.obj(
+  var stream = composer.pipeline(
     prepare(folderResolver, optResolver),
-    mkdirpStream.obj(dirpath),
+    mkdirpStream(dirpath),
     linkFile(optResolver)
   );
 
