@@ -3,7 +3,7 @@
 var os = require('os');
 var path = require('path');
 
-var through = require('through2');
+var Transform = require('streamx').Transform;
 
 var fo = require('../file-operations');
 
@@ -11,7 +11,7 @@ var isWindows = (os.platform() === 'win32');
 
 function linkStream(optResolver) {
 
-  function linkFile(file, enc, callback) {
+  function linkFile(file, callback) {
     var isRelative = optResolver.resolve('relativeSymlinks', file);
     var flags = fo.getFlags({
       overwrite: optResolver.resolve('overwrite', file),
@@ -83,7 +83,9 @@ function linkStream(optResolver) {
     }
   }
 
-  return through.obj(linkFile);
+  return new Transform({
+    transform: linkFile,
+  });
 }
 
 module.exports = linkStream;
