@@ -4,14 +4,14 @@ var path = require('path');
 
 var fs = require('graceful-fs');
 var Vinyl = require('vinyl');
-var through = require('through2');
+var Transform = require('streamx').Transform;
 
 function prepareSymlink(folderResolver, optResolver) {
   if (!folderResolver) {
     throw new Error('Invalid output folder');
   }
 
-  function normalize(file, enc, cb) {
+  function normalize(file, cb) {
     if (!Vinyl.isVinyl(file)) {
       return cb(new Error('Received a non-Vinyl object in `symlink()`'));
     }
@@ -45,7 +45,9 @@ function prepareSymlink(folderResolver, optResolver) {
     cb(null, file);
   }
 
-  return through.obj(normalize);
+  return new Transform({
+    transform: normalize,
+  });
 }
 
 module.exports = prepareSymlink;
