@@ -29,7 +29,8 @@ var leBomInputPath = path.join(inputBase, './bom-utf16le.txt');
 var bomContents = 'This file is saved as UTF-X with the appropriate BOM.\n';
 var beNotBomInputPath = path.join(inputBase, './not-bom-utf16be.txt');
 var leNotBomInputPath = path.join(inputBase, './not-bom-utf16le.txt');
-var notBomContents = 'This file is saved as UTF-16-X. It contains some garbage at the start that looks like a UTF-8-encoded BOM (but isn\'t).\n';
+var notBomContents =
+  "This file is saved as UTF-16-X. It contains some garbage at the start that looks like a UTF-8-encoded BOM (but isn't).\n";
 var ranBomInputPath = path.join(inputBase, './ranbom.bin');
 // Used for encoding tests
 var encodedInputPath = path.join(inputBase, './enc-gb2312.txt');
@@ -38,8 +39,14 @@ var encodedContents = '\u5b54\u5b50\u8bf4\u590d\u6d3b\u8282\u5f69\u86cb\n';
 var symlinkNestedTarget = path.join(inputBase, './foo/bar/baz.txt');
 var symlinkPath = path.join(outputBase, './test-symlink');
 var symlinkDirpath = path.join(outputBase, './test-symlink-dir');
-var symlinkMultiDirpath = path.join(outputBase, './test-multi-layer-symlink-dir');
-var symlinkMultiDirpathSecond = path.join(outputBase, './test-multi-layer-symlink-dir2');
+var symlinkMultiDirpath = path.join(
+  outputBase,
+  './test-multi-layer-symlink-dir'
+);
+var symlinkMultiDirpathSecond = path.join(
+  outputBase,
+  './test-multi-layer-symlink-dir2'
+);
 var symlinkNestedFirst = path.join(outputBase, './test-multi-layer-symlink');
 var symlinkNestedSecond = path.join(outputBase, './foo/baz-link.txt');
 // Paths that don't exist
@@ -49,7 +56,8 @@ var neInputDirpath = path.join(neInputBase, './foo');
 var neOutputDirpath = path.join(neOutputBase, './foo');
 // Used for contents of files
 var contents = 'Hello World!\n';
-var sourcemapContents = '//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9maXh0dXJlcyIsIm5hbWVzIjpbXSwibWFwcGluZ3MiOiIiLCJzb3VyY2VzIjpbIi4vZml4dHVyZXMiXSwic291cmNlc0NvbnRlbnQiOlsiSGVsbG8gV29ybGQhXG4iXX0=';
+var sourcemapContents =
+  '//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9maXh0dXJlcyIsIm5hbWVzIjpbXSwibWFwcGluZ3MiOiIiLCJzb3VyY2VzIjpbIi4vZml4dHVyZXMiXSwic291cmNlc0NvbnRlbnQiOlsiSGVsbG8gV29ybGQhXG4iXX0=';
 
 module.exports = {
   inputRelative: inputRelative,
