@@ -1,75 +1,191 @@
 'use strict';
 
-var miss = require('mississippi');
 var expect = require('expect');
 
-var to = miss.to;
-var from = miss.from;
-var through = miss.through;
+module.exports = function makeUtils(stream) {
+  function chunks(length) {
+    var arr = new Array(length);
+    arr.fill('a');
+    return stream.Readable.from(arr);
+  }
 
-function string(length) {
-  return from(function(size, next) {
-    if (length <= 0) {
-      next(null, null);
-      return;
-    }
+  function rename(filepath) {
+    return new stream.Transform({
+      objectMode: true,
+      transform: function (file, enc, cb) {
+        if (typeof enc === 'function') {
+          cb = enc;
+        }
 
-    var chunkSize = size <= length ? size : length;
+        file.path = filepath;
+        cb(null, file);
+      },
+    });
+  }
 
-    length -= size;
+  function includes(obj) {
+    return new stream.Transform({
+      objectMode: true,
+      transform: function (file, enc, cb) {
+        if (typeof enc === 'function') {
+          cb = enc;
+        }
 
-    var chunk = '';
-    for (var x = 0; x < chunkSize; x++) {
-      chunk += 'a';
-    }
+        expect(file.path).toEqual(obj.path);
+        cb(null, file);
+      },
+    });
+  }
 
-    next(null, chunk);
-  });
-}
+  function count(value) {
+    var count = 0;
+    return new stream.Transform({
+      objectMode: true,
+      transform: function (file, enc, cb) {
+        if (typeof enc === 'function') {
+          cb = enc;
+        }
 
-function rename(filepath) {
-  return through.obj(function(file, enc, cb) {
-    file.path = filepath;
-    cb(null, file);
-  });
-}
+        count++;
+        cb(null, file);
+      },
+      flush: function (cb) {
+        expect(count).toEqual(value);
+        cb();
+      },
+    });
+  }
 
-function includes(obj) {
-  return through.obj(function(file, enc, cb) {
-    expect(file).toInclude(obj);
-    cb(null, file);
-  });
-}
+  function slowCount(value) {
+    var count = 0;
+    return new stream.Writable({
+      objectMode: true,
+      write: function (file, enc, cb) {
+        if (typeof enc === 'function') {
+          cb = enc;
+        }
 
-function count(value) {
-  var count = 0;
-  return through.obj(function(file, enc, cb) {
-    count++;
-    cb(null, file);
-  }, function(cb) {
-    expect(count).toEqual(value);
-    cb();
-  });
-}
+        count++;
 
-function slowCount(value) {
-  var count = 0;
-  return to.obj(function(file, enc, cb) {
-    count++;
+        setTimeout(function () {
+          cb(null, file);
+        }, 250);
+      },
+      end: function (data, enc, cb) {
+        expect(count).toEqual(value);
+        cb();
+      },
+    });
+  }
 
-    setTimeout(function() {
-      cb(null, file);
-    }, 250);
-  }, function(cb) {
-    expect(count).toEqual(value);
-    cb();
-  });
-}
+  function concatString(fn, timeout) {
+    var collect = '';
+    return new stream.Writable({
+      objectMode: true,
+      write: function (chunk, enc, cb) {
+        if (typeof enc === 'function') {
+          cb = enc;
+        }
+        setTimeout(function () {
+          collect = collect + chunk;
+          cb();
+        }, timeout || 1);
+      },
+      final: function (cb) {
+        if (typeof fn === 'function') {
+          fn(collect);
+        }
 
-module.exports = {
-  string: string,
-  rename: rename,
-  includes: includes,
-  count: count,
-  slowCount: slowCount,
+        cb();
+      },
+    });
+  }
+
+  function concatBuffer(fn, timeout) {
+    var collect = Buffer.alloc(0);
+    return new stream.Writable({
+      objectMode: true,
+      write: function (chunk, enc, cb) {
+        if (typeof enc === 'function') {
+          cb = enc;
+        }
+        setTimeout(function () {
+          collect = Buffer.concat([collect, chunk]);
+          cb();
+        }, timeout || 1);
+      },
+      final: function (cb) {
+        if (typeof fn === 'function') {
+          fn(collect);
+        }
+
+        cb();
+      },
+    });
+  }
+
+  function concatArray(fn, timeout) {
+    var collect = [];
+    return new stream.Writable({
+      objectMode: true,
+      write: function (chunk, enc, cb) {
+        if (typeof enc === 'function') {
+          cb = enc;
+        }
+        setTimeout(function () {
+          collect.push(chunk);
+          cb();
+        }, timeout || 1);
+      },
+      final: function (cb) {
+        if (typeof fn === 'function') {
+          fn(collect);
+        }
+
+        cb();
+      },
+    });
+  }
+
+  function compareContents(comparator) {
+    return new stream.Transform({
+      objectMode: true,
+      transform: function (file, enc, cb) {
+        if (typeof enc === 'function') {
+          cb = enc;
+        }
+        stream.pipeline(
+          [file.contents, concatBuffer(comparator)],
+          function (err) {
+            cb(err, file);
+          }
+        );
+      },
+    });
+  }
+
+  function fromError(error) {
+    return new stream.Readable({
+      read: function (cb) {
+        if (typeof cb === 'function') {
+          cb(error);
+        } else {
+          this.destroy(error);
+        }
+      },
+    });
+  }
+
+  return {
+    chunks: chunks,
+    rename: rename,
+    includes: includes,
+    count: count,
+    slowCount: slowCount,
+    concatString: concatString,
+    concatBuffer: concatBuffer,
+    concatArray: concatArray,
+    compareContents: compareContents,
+    fromError: fromError,
+  };
 };
